Filter feature is working, added some basic css
