+++ src/models/User.js
@@ -15,6 +15,18 @@ const UserSchema = mongoose.Schema({
         type: String,
         required: true,
     },
+    topGame: {
+        type: String,
+        required: true,
+    },
+    time: {
+        type: String,
+        required: true,
+    },
+    playstyle: {
+        type: String,
+        required: true,
+    },
 });
 
 const User = mongoose.model("User", UserSchema);
