--- src/views/pages/filter/chooseGame.njk
@@ -1,11 +1,0 @@
-{% extends "layouts/layout.njk" %}
-{% block title %}
-    Games
-{% endblock %}
-{% block content %}
-    <div class="games">
-            {% for game in gameList %}
-                <a href="/filter/{{ game }}">{{ game }}</a>
-            {% endfor %}
-    </div>
-{% endblock %}
