+++ static/public/css/style.css
@@ -1,23 +1,32 @@
 /* Imports */
+
+
 /* Variables */
+
 .hidden {
-  display: none; }
+    display: none;
+}
 
 .chatList form {
-  display: flex;
-  flex-direction: column; }
-  .chatList form button {
+    display: flex;
+    flex-direction: column;
+}
+
+.chatList form button {
     width: 100%;
     height: 50px;
     margin: 0;
     padding: 0;
     background-color: white;
     border: none;
-    border-bottom: solid black 1px; }
+    border-bottom: solid black 1px;
+}
 
 .chatSelf {
-  margin-top: 3.5em; }
-  .chatSelf .chatHeader {
+    margin-top: 3.5em;
+}
+
+.chatSelf .chatHeader {
     background-color: grey;
     position: fixed;
     top: 3em;
@@ -26,13 +35,19 @@
     color: white;
     display: flex;
     flex-direction: row;
-    padding: 0.3em; }
-    .chatSelf .chatHeader a {
-      text-decoration: none;
-      color: white; }
-    .chatSelf .chatHeader div {
-      margin-left: 1em; }
-  .chatSelf #form {
+    padding: 0.3em;
+}
+
+.chatSelf .chatHeader a {
+    text-decoration: none;
+    color: white;
+}
+
+.chatSelf .chatHeader div {
+    margin-left: 1em;
+}
+
+.chatSelf #form {
     background: rgba(0, 0, 0, 0.15);
     padding: 0.25rem;
     position: fixed;
@@ -42,29 +57,39 @@
     display: flex;
     height: 3rem;
     box-sizing: border-box;
-    backdrop-filter: blur(10px); }
-  .chatSelf #input {
+    backdrop-filter: blur(10px);
+}
+
+.chatSelf #input {
     border: none;
     padding: 0 1rem;
     flex-grow: 1;
     border-radius: 2rem;
-    margin: 0.25rem; }
-  .chatSelf #input:focus {
-    outline: none; }
-  .chatSelf #form > button {
+    margin: 0.25rem;
+}
+
+.chatSelf #input:focus {
+    outline: none;
+}
+
+.chatSelf #form>button {
     background: #333;
     border: none;
     padding: 0 1rem;
     margin: 0.25rem;
     border-radius: 3px;
     outline: none;
-    color: #fff; }
-  .chatSelf #messages {
+    color: #fff;
+}
+
+.chatSelf #messages {
     list-style-type: none;
     margin: 0;
     margin-bottom: 2em;
-    padding: 0; }
-  .chatSelf #messages > li {
+    padding: 0;
+}
+
+.chatSelf #messages>li {
     padding: 0.5rem 1rem;
     margin-bottom: 0.3em;
     width: fit-content;
@@ -75,7 +100,73 @@
     height: fit-content;
     display: flex;
     word-wrap: break-word;
-    word-break: break-all; }
-  .chatSelf .messageSelf {
+    word-break: break-all;
+}
+
+.chatSelf .messageSelf {
     margin-left: auto;
-    background-color: #343480 !important; }
+    background-color: #343480 !important;
+}
+
+
+/* --------------Filter-------------- */
+
+.filter form {
+    margin-top: 30px;
+}
+
+.filter form fieldset {
+    margin-bottom: 30px;
+    border: none;
+    padding-left: 0;
+    margin-left: 0;
+}
+
+.filter form fieldset select {
+    padding: 10px;
+    width: 40%;
+    margin-top: 10px;
+}
+
+.filter form fieldset:nth-of-type(2),
+.filter form fieldset:nth-of-type(3) {
+    display: flex;
+    flex-direction: column;
+}
+
+.filter form input[type="submit"] {
+    padding: 5px 15px;
+}
+
+
+/* --------------Matches-------------- */
+
+.matches>section {
+    border: 1px solid;
+    display: flex;
+    flex-wrap: wrap;
+    width: 80%;
+    padding: 10px;
+    margin: 20px auto;
+    align-items: center;
+    justify-content: space-around;
+}
+
+.matches>section p {
+    margin: 0 0 5px;
+}
+
+.matches>section p:nth-of-type(1) {
+    flex-basis: 100%;
+    margin: 5px 0px 10px;
+    font-size: 20px;
+    font-weight: bold;
+}
+
+.matches>section a {
+    display: block;
+    text-decoration: none;
+    color: black;
+    border: 1px solid;
+    padding: 8px 18px;
+}
