+++ static/public/scss/style.scss
@@ -1,5 +1,8 @@
 @charset "UTF-8";
 
 /* Imports */
-@import "1-settings/_variables.scss", "chat/_chatList.scss",
-  "chat/_chatSelf.scss";
+
+@import "1-settings/_variables.scss",
+"chat/_chatList.scss",
+"chat/_chatSelf.scss",
+"chat/_filter.scss";
